Restyle current activity page
