+++ package.json
@@ -27,6 +27,7 @@
     "react": "^15.6.2",
     "react-dom": "^15.6.2",
     "react-modal": "^3.1.4",
+    "react-rangeslider": "^2.2.0",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "react-slick": "^0.23.1",
